Separate shipment into a unique model
